+++ CHANGELOG.md
@@ -17,6 +17,11 @@
     Click to see more.
   </summary>
 
+### :bug: Bug Fix
+
+- `hashtag`
+  - [#2796](https://github.com/wix/ricos/pull/2796) disable hashtag inside quotes for avoid hashtag in code-snippet selectors
+
 </details>
 <hr/>